+++ test-workspace/config/spa/vendor-all.js
@@ -2,5 +2,5 @@ module.exports = {
   entry: {
     main: './spa/scripts/vendor-main.js'
   },
-  splitAllVendors: true
+  extractAllVendors: true
 };